Update VC project configurations, eliminate some x64 warnings
Former-commit-id: 153682f7c60c26d1124d907e0f05bba471d628ba

+++ test/test_util.h
@@ -206,7 +206,7 @@ public:
      */
     int ParsedArgc()
     {
-        return keys.size();
+        return (int) keys.size();
     }
 
     /**
